Convert metabase/hoc files to TypeScript
